Update and rename file
